+++ src/demo/scss/components/load-more.scss
@@ -1,0 +1,3 @@
+#loadMore {
+    @apply .button
+}
