+++ src/scss/components/load-more.scss
@@ -1,0 +1,4 @@
+#loadMore {
+    padding: .5rem 2rem .5rem 2rem;
+    border: 1px rgba(226, 232, 240, var(--border-opacity)) solid;
+}
